\name{plotAdelaideFD}
\alias{plotAdelaideFD}
\title{
plotAdelaideFD
}
\description{
Plot data generated by fitAdelaideFD.
}
\arguments{
  \item{fd_demand}{
List of data made by fitAdelaideFD to plot.
}
}

\usage{
plotAdelaideFD(fd_demand)
}
\author{
Cristina Anton and Iain Smith
}

\examples{
data <- fitAdelaideFD()
plotAdelaideFD(data)
}
