\name{plotRegFD}
\alias{plotRegFD}
\title{
plotRegFD
}
\description{
Plot data generated by  genFromModelRegFD.
}
\arguments{
  \item{gen_out}{
List of data made by  genFromModelRegFD to plot.
}
}

\usage{
plotRegFD(model_reg)
}
\author{
Cristina Anton and Iain Smith
}

\examples{
res_reg=readRDS("input2n.RData")
model_reg <- genFromModelRegFD(res_reg, ncurves = c(300, 300),mu_div = 1.9)
plotRegFD(model_reg)
}
